+++ packages/expo-clipboard/src/Clipboard.ts
@@ -1,4 +1,4 @@
-import { EventEmitter, Subscription, UnavailabilityError } from 'expo-modules-core';
+import { EventEmitter, Subscription, UnavailabilityError, Platform } from 'expo-modules-core';
 
 import ExpoClipboard from './ExpoClipboard';
 
@@ -33,14 +33,31 @@ export async function getStringAsync(): Promise<string> {
  *
  * @param text The string to save to the clipboard.
  *
+ * @returns On web, this returns a promise that fulfills to a boolean value indicating whether or not
+ * the string was saved to the user's clipboard. On iOS and Android, the promise always resolves to `true`.
+ */
+export async function setStringAsync(text: string): Promise<boolean> {
+  if (!ExpoClipboard.setStringAsync) {
+    throw new UnavailabilityError('Clipboard', 'setStringAsync');
+  }
+  return ExpoClipboard.setStringAsync(text);
+}
+
+/**
+ * Sets the content of the user's clipboard.
+ * @deprecated Deprecated. Use [`setStringAsync()`](#setstringasynctext) instead.
+ *
  * @returns On web, this returns a boolean value indicating whether or not the string was saved to
  * the user's clipboard. On iOS and Android, nothing is returned.
  */
 export function setString(text: string): void {
-  if (!ExpoClipboard.setString) {
-    throw new UnavailabilityError('Clipboard', 'setString');
+  if (Platform.OS === 'web') {
+    // on web, we need to return legacy method,
+    // because of different return type
+    return ExpoClipboard.setString(text);
+  } else {
+    setStringAsync(text);
   }
-  return ExpoClipboard.setString(text);
 }
 
 /**