+++ packages/expo-clipboard/CHANGELOG.md
@@ -4,7 +4,11 @@
 
 ### 🛠 Breaking changes
 
+- Deprecated `setString`. Use `setStringAsync` instead. ([#16320](https://github.com/expo/expo/pull/16320) by [@barthap](https://github.com/barthap))
+
 ### 🎉 New features
+
+- Native module on Android is now written in Kotlin using the new API. ([#16269](https://github.com/expo/expo/pull/16269) by [@barthap](https://github.com/barthap))
 
 ### 🐛 Bug fixes